requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-allauth==0.45.0
django-filter==21.1
django-ipware==4.0.0
django-oauth-toolkit==1.5.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-structlog==2.1.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt
djoser
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
pymemcache==3.5.0
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
simplejson==3.17.5
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
structlog==21.1.0
uritemplate==3.0.1
urllib3==1.26.7
wrapt==1.12.1
